Update Cargo files and `gitignore`
This commit will add in the `Cargo.toml` and `Cargo.lock` files. It will
also update the `.gitignore` file to include `Cargo.lock`, which is
normally disabled by default on Rust version control systems. This is
because Metis is intended to be run as a binary rather than a library.